Kotlin
I want to like nim, but the standard library feels messy and
undiscoverable. I like the significant whitespace, and I like the
optional parameters and the object dotting behavior, but I'm not a fan
of typing the anonymous blocks and I really miss having an IDE.

Kotlin adds a little bit of weight back to the code, with all of the
symbols, but I prefer it's standard library having spent most of my time
in a java environment.

Furthermore, Day 3 was complex enough that I was starting to miss my
comfortable coding environments. I'm at least now doing the problems in
a language I don't get to use everyday.

And yes, Justin, if you're reading this, I do think of you everytime
I rewrite code in a different language.